수정

+++ project_test/Libraries/ALARM/alarm.c
@@ -24,7 +24,7 @@ void Alarm_Init(void) {
 
 void Alarm_Start(uint16_t minutes) {
     if (minutes > 0) {
-        *p_countdown_seconds = minutes * 60;
+        *p_countdown_seconds = (uint32_t)minutes * 60;
         *p_alarm_state = STATE_COUNTDOWN;
         *p_elapsed_seconds = 0;
         LCD_Clear(WHITE);
@@ -39,8 +39,8 @@ void Alarm_Process(void) {
 
     switch (*p_alarm_state) {
         case STATE_COUNTDOWN:
-            if (last_displayed_second != *p_countdown_seconds) {
-                last_displayed_second = *p_countdown_seconds;
+            if (last_displayed_second != (int32_t)*p_countdown_seconds) {
+                last_displayed_second = (int32_t)*p_countdown_seconds;
                 sprintf(lcd_buffer, "남은 시간: %02d분 %02d초", (int)(*p_countdown_seconds / 60), (int)(*p_countdown_seconds % 60));
                 LCD_ShowString(40, 130, (u8*)lcd_buffer, BLUE, WHITE);
             }
@@ -68,7 +68,7 @@ void Alarm_Reset(void) {
     *p_elapsed_seconds = 0;
     *p_countdown_seconds = 0;
     TIM_Cmd(TIM2, DISABLE);
-    BUZZER_Off(); // PA8 핀 High로 설정하여 부저 정지
+    BUZZER_Off(); 
     LCD_Clear(WHITE);
     LCD_ShowString(40, 100, (u8*)"알람 대기중...", BLUE, WHITE);
 }
